🔧 chore: Update iOS app store link to use numeric ID
This commit fine-tunes the redirection logic for iOS devices by replacing the placeholder for the app's bundle ID with the app's numeric ID in the store link. This adjustment ensures that the redirection URL conforms to the standard URL scheme required by the Apple App Store, thereby improving the reliability of the redirection process for iOS users.

Changes made:
- Added a new environment variable `IOS_APP_NUMERIC_ID` to store the numeric ID of the iOS application. This ID is crucial for constructing the correct URL to the app's page on the Apple App Store.
- Modified the `redirectPage` function to use `iOSAppNumericId` for generating the store link for iOS devices. The store link now accurately reflects the format `https://apps.apple.com/app/id${iOSAppNumericId}`, aligning with Apple's URL scheme for app store listings.

This update enhances the user experience for iOS users by ensuring that they are directed to the correct app store page, facilitating easier access to the app. By leveraging the numeric ID, the redirection link becomes more precise and universally applicable, regardless of potential variations in bundle IDs or app names.